demo/requirements.txt
# Demo requirements
streamlit>=1.22.0
requests>=2.28.0
Pillow>=9.5.0